// embedded-components/src/api/generated/smbdo.schemas.ts
export interface QuestionListResponse {
  metadata?: PageMetaData;
  /**
   * @minItems 0
   * @maxItems 100
   */
  questions?: QuestionResponse[];
}

/**
 * Describes the answer value that would require dependent questions.
 * @maxLength 255
 */
export type AnyValuesMatch = string;

export type QuestionResponseSubQuestionsItem = {
  anyValuesMatch?: AnyValuesMatch;
  questionIds?: QuestionIdList;
};

export type ResponseSchemaType =
  (typeof ResponseSchemaType)[keyof typeof ResponseSchemaType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ResponseSchemaType = {
  array: 'array',
} as const;

/**
 * A subset of JSON Schema used to validate the response values.
 */
export interface ResponseSchema {
  items?: ResponseSchemaItem;
  maxItems?: number;
  minItems?: number;
  type?: ResponseSchemaType;
}

export type ResponseSchemaItemType =
  (typeof ResponseSchemaItemType)[keyof typeof ResponseSchemaItemType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ResponseSchemaItemType = {
  boolean: 'boolean',
  string: 'string',
  number: 'number',
  integer: 'integer',
  enum: 'enum',
} as const;

/**
 * Only applicable to string, number, and integer.
 */
export type ResponseSchemaItemFormat =
  (typeof ResponseSchemaItemFormat)[keyof typeof ResponseSchemaItemFormat];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ResponseSchemaItemFormat = {
  float: 'float',
  double: 'double',
  int32: 'int32',
  int64: 'int64',
  date: 'date',
} as const;

/**
 * A limited subset of JSON Schema used to validate the response value items.
 */
export interface ResponseSchemaItem {
  exclusiveMaximum?: boolean;
  exclusiveMinimum?: boolean;
  /** Only applicable to string, number, and integer. */
  format?: ResponseSchemaItemFormat;
  maximum?: number;
  maxLength?: number;
  minimum?: number;
  minLength?: number;
  /** Only applicable to string. */
  pattern?: string;
  type?: ResponseSchemaItemType;
}

/**
 * A description of the question.
 */
export type QuestionDescription = string;

/**
 * The identifier for a language and location.
 */
export type LocaleId = string;

/**
 * Content displayed to a user for a given locale.
 */
export interface ContentItem {
  description?: string;
  /** Exact text to be displayed to a user. */
  label: string;
  locale: LocaleId;
}

export interface QuestionResponse {
  /**
   * @minItems 1
   * @maxItems 50
   */
  content?: ContentItem[];
  defaultLocale?: LocaleId;
  description?: QuestionDescription;
  id?: QuestionId;
  parentQuestionId?: QuestionId;
  responseSchema?: ResponseSchema;
  /** @minItems 0 */
  subQuestions?: QuestionResponseSubQuestionsItem[];
}
// ...
export type QuestionId = string;

/**
 * A list of question IDs.
 * @minItems 0
 * @maxItems 200
 */
export type QuestionIdList = QuestionId[];
// ...
export interface PageMetaData {
  /**
   * Number of records per page.

   * @minimum 1
   * @maximum 25
   */
  limit?: number;
  /**
   * Page number.

   * @minimum 0
   * @maximum 2147483647
   */
  page?: number;
  /**
   * Total number of items.

   * @minimum 0
   * @maximum 9999999999
   */
  total?: number;
}